Added a few more default extensions.

+++ pull.sh
@@ -1,17 +1,27 @@
 mkdir pull
 cd pull
 
-PKGS=$(ls ../.cabal-sandbox/lib/i386-linux-ghc-7.8.3)
+# TODO get a consistent snapshot somewhere, extracting from local cabal sandbox is an approximate hack
+
+echo ==== pkgs
+PKGS=$(ls ../.cabal-sandbox/lib/i386-linux-ghc-7.8.3 /usr/local/lib/ghc-7.8.3 | grep -v ':' | grep '\-[0-9\.]\+' | sort)
+[ -f list.txt ] && list.txt
 for PKG in $PKGS
 do
   LINK=https://hackage.haskell.org/package/$PKG/$PKG.tar.gz
-  wget $LINK
+  echo $LINK >> list.txt
 done
+cat list.txt
 
+echo ==== wget
+wget -i list.txt 2> wget.log
+
+echo ==== unpack
 ls *tar.gz | xargs -n1 tar xzf
 
+echo ==== list
 SRCS=$(find . | grep '\.hs$')
-rm files.txt
+[ -f files.txt ] && rm files.txt
 for SRC in $SRCS
 do
   FULL_PATH=$(readlink -m $SRC)
